refactor(oracle): migrate ipv4 to per-protocol plugins
