+++ tools/oracle/engine/generator.py
@@ -29,6 +29,7 @@ from .sampling import (
     _integer_domain_value,
     _IPV4_DOCUMENTATION_NETWORKS,
     _IPV6_DOCUMENTATION_NETWORK,
+    _is_ipv4_root_dhcp_stack,
     _mac_for_domain,
     _next_layer_after,
     bounded_int,
@@ -263,17 +264,6 @@ _SUPPORTED_FIELDS: dict[str, set[str]] = {
         "flags",
         "message_id",
         "length",
-    },
-    "ipv4": {
-        "ds_field",
-        "src",
-        "dst",
-        "ttl",
-        "protocol",
-        "identification",
-        "flags",
-        "fragment_offset",
-        "options",
     },
     "ipv6": {"src", "dst", "traffic_class", "flow_label", "next_header", "hop_limit"},
     # OSPFv2 common-header fields declared in specs/layers/ospf.yaml. The
@@ -1815,10 +1805,6 @@ class PacketGenerator:
                 case=case,
                 behavior=behavior,
             )
-        elif feature == "ipv4_options" and "ipv4" in fields:
-            fields["ipv4"]["options"] = {"hex": _ipv4_options_hex(case, behavior)}
-            fields["ipv4"]["flags"] = "none"
-            fields["ipv4"]["fragment_offset"] = 0
         elif feature == "ipv6_fragment_routing":
             if "ipv6_fragment" in fields:
                 fields["ipv6_fragment"]["more_fragments"] = False
@@ -2380,8 +2366,6 @@ class PacketGenerator:
                 field_spec=field_spec,
                 current_fields=current_fields,
             )
-        if layer == "ipv4":
-            return _sample_ipv4_field(ctx, field_name, domain, current_fields)
         if layer == "ipv6":
             return _sample_ipv6_field(ctx, field_name, domain)
         if layer == "udp":
@@ -2551,89 +2535,6 @@ def _is_boundary_domain(domain: object) -> bool:
         "zero",
         "zero_ipv4",
     }
-
-
-def _is_ipv4_root_dhcp_stack(stack: Sequence[str]) -> bool:
-    """Return True for the IPv4-root, unicast live DHCP stack.
-
-    The ``ipv4 / udp / dhcp`` stack carries DHCP as a one-way unicast oracle
-    packet between provider endpoints. It has no Ethernet frame, so link-layer
-    broadcast delivery, the IPv4 limited-broadcast destination, and the DHCP
-    broadcast flag have no meaning and would make the packet ineligible for
-    provider-backed live exchange. The Ethernet-root DHCP stack keeps those
-    domains for offline link-layer coverage.
-    """
-
-    return "dhcp" in stack and "ethernet" not in stack
-
-
-def _ipv4_for_domain(ctx: _SamplingContext, domain: object, default: str, *, dst: bool) -> str:
-    if domain == "zero":
-        return "0.0.0.0"
-    if domain == "broadcast" and dst:
-        if _is_ipv4_root_dhcp_stack(ctx.stack):
-            return default
-        return "255.255.255.255"
-    return default
-
-
-def _sample_ipv4_field(
-    ctx: _SamplingContext,
-    field_name: str,
-    domain: object,
-    current_fields: Mapping[str, object],
-) -> object:
-    if field_name == "src":
-        return _ipv4_for_domain(ctx, domain, ctx.src_ipv4, dst=False)
-    if field_name == "dst":
-        return _ipv4_for_domain(ctx, domain, ctx.dst_ipv4, dst=True)
-    if field_name == "ds_field":
-        if "boundary-fields" in ctx.case:
-            return 0xFF
-        if domain == "dscp_ef_ecn_not_ect":
-            return 0b10111000
-        if domain == "dscp_cs5_ecn_ect0":
-            return 0b10100010
-        return _integer_domain_value(ctx, domain, field_name, bits=8)
-    if field_name == "ttl":
-        if "ttl-255" in ctx.case:
-            return 255
-        return _integer_domain_value(ctx, domain, field_name, bits=8)
-    if field_name == "protocol":
-        return _ipv4_protocol_for_stack(ctx.stack)
-    if field_name == "identification":
-        return _integer_domain_value(ctx, domain, field_name, bits=16)
-    if field_name == "flags":
-        if "fragment-mf-offset" in ctx.case:
-            return "mf"
-        if any(layer in ctx.stack for layer in ("tcp", "udp", "icmp", "dns", "dhcp")):
-            return "none"
-        # ESP/AH/IKEv2 datagrams are never carried as IP fragments in the oracle
-        # corpus: libcrafter (correctly) leaves a non-first fragment body opaque
-        # and does not dispatch the inner protocol on a fragment, while Scapy
-        # binds the IP protocol number regardless of fragmentation. Pin the outer
-        # (and inner tunnel) IP to non-fragmented so both backends dissect the
-        # ESP/AH/IKEv2 layer and the decoded models agree.
-        if any(layer in ctx.stack for layer in ("esp", "ah", "ikev2")):
-            return "none"
-        if ctx.profile == "smoke":
-            return "none"
-        return str(domain)
-    if field_name == "fragment_offset":
-        if "fragment-mf-offset" in ctx.case:
-            return 1
-        if any(layer in ctx.stack for layer in ("tcp", "udp", "icmp", "dns", "dhcp")):
-            return 0
-        if any(layer in ctx.stack for layer in ("esp", "ah", "ikev2")):
-            return 0
-        if domain == "non_initial":
-            return 1
-        if current_fields.get("flags") == "mf":
-            return _integer_domain_value(ctx, domain, field_name, bits=13)
-        return 0
-    if field_name == "options":
-        return {"hex": _ipv4_options_hex(ctx.case, str(domain))}
-    raise ValueError(f"spec error: unsupported ipv4 field sampler: {field_name}")
 
 
 def _sample_ipv6_field(ctx: _SamplingContext, field_name: str, domain: object) -> object:
@@ -3175,19 +3076,6 @@ def _tcp_option_case_hex(behavior: str) -> str:
     if key not in _TCP_OPTION_CASE_HEX:
         raise ValueError(f"spec error: no tcp option hex for behavior {behavior!r}")
     return _TCP_OPTION_CASE_HEX[key]
-
-
-def _ipv4_options_hex(case: str, behavior: str) -> str:
-    key = f"{case} {behavior}".replace("_", "-")
-    if "source-route" in key:
-        return "830704c0000201"
-    if "record-route" in key:
-        return "07070400000000"
-    if "timestamp" in key or "traceroute" in key or "generic" in key:
-        return "440c05000000000000000000120c00010000ffffc0000201"
-    if "nop" in key:
-        return "01010101"
-    return "00000000"
 
 
 def _icmp_error_type_for_case(case: str, behavior: str, *, ipv6: bool) -> str:
@@ -5971,13 +5859,6 @@ def _udp_option_cases_for_stack(stack: Sequence[str], cases: Sequence[str]) -> l
             continue
         output.append(case)
     return output
-
-
-def _ipv4_protocol_for_stack(stack: Sequence[str]) -> str:
-    next_layer = _next_layer_after(stack, "ipv4")
-    if next_layer in {"icmp", "tcp", "udp"}:
-        return next_layer
-    return "unknown"
 
 
 def _ipv6_next_header_for_stack(stack: Sequence[str], layer: str) -> str:
